Rename infraestructure directories
