+++ src/main/java/com/example/soccer/score/infrastructure/InMemoryStandingRepository.java
@@ -1,4 +1,4 @@
-package com.example.soccer.score.infraestructure;
+package com.example.soccer.score.infrastructure;
 
 import com.example.soccer.score.domain.Standing;
 import com.example.soccer.score.domain.StandingRepository;
