+++ src/main/java/com/example/soccer/score/infrastructure/ScoreController.java
@@ -1,4 +1,4 @@
-package com.example.soccer.score.infraestructure;
+package com.example.soccer.score.infrastructure;
 
 import com.example.soccer.score.application.GetStandingsQuery;
 import com.example.soccer.score.application.GetTopScorersQuery;
